Use ESBuild to produce a small log library
