+++ package.json
@@ -2,9 +2,12 @@
   "name": "homey-log",
   "version": "3.0.0",
   "description": "Interface with Sentry for Homey",
-  "main": "index.js",
+  "main": "./build/index.js",
+  "module": "./build/index.mjs",
+  "types": "./build/index.d.ts",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
+    "build": "node build.js",
     "lint": "eslint .",
     "jsdoc": "jsdoc --configure ./jsdoc.json",
     "jsdoc:serve": "serve ./jsdoc",
@@ -25,13 +28,12 @@
     "node": ">=16"
   },
   "homepage": "https://github.com/athombv/node-homey-log#readme",
-  "dependencies": {
-    "@sentry/node": "^8.33.1"
-  },
   "devDependencies": {
     "@athombv/jsdoc-template": "github:athombv/homey-jsdoc-template#v1.6.1",
+    "@sentry/node": "^8.33.1",
     "@types/homey": "npm:homey-apps-sdk-v3-types@^0.3.8",
     "@types/node": "^16.18.0",
+    "esbuild": "^0.24.0",
     "eslint": "^7.32.0",
     "eslint-config-athom": "^2.1.1",
     "homey": "^3.7.0",
@@ -42,7 +44,6 @@
     "typescript": "^4.7.4",
     "watch": "^1.0.2"
   },
-  "types": "types/index.d.ts",
   "watch": {
     "jsdoc": {
       "patterns": [
